backend/requirements.txt
# requirements.txt - Required dependencies
"""
Flask
flask-cors
google-generativeai
mysql-connector-python
"""